Update function for apc diagnosis
To include diagnosis in any position in the function. Previously we just included primary diagnosis and first code of secondary diagnosis.

The definition of Covid-19 hospitalisation `sub_date_covid19_hospital` did not use any created function as its definition is very unique and not worth creating a function for it self.